Card selection, multiple partners
I added the option to select a partner, and more than 1 if needed. Although I haven't quite set up the ability to send more than 1 partner to the server.

I also finished enabling the winner of the bid to select the cards to send back to the server and then update his hand.