Fixed some bugs that showed up inbetween rounds
Going between rounds seems to work properly now. There is also now a visual indicator when it is your turn to bid or play a card.

Winning the game is also now detected. But it is utilising alert, and this causes some problems when trying to start a new game and not everyone has clicked ok on the alert.