Changed the endgame behavior a bit
No longer uses alert to announce game winner.

Players are given 15 seconds to look at the final scores. We need to change this so that all scores are displayed.

Also, we should probably check for ties.

Then there seems to be a bit of a glitch when sending players back into the lobby.